+++ R/specs.R
@@ -26,9 +26,9 @@
 #'   K600_daily_meanlog_sdlog, K600_daily_sdlog}. \item If 
 #'   \code{pool_K600=='linear'} then \code{lnK600_lnQ_intercept_mu, 
 #'   lnK600_lnQ_intercept_sigma, lnK600_lnQ_slope_mu, lnK600_lnQ_slope_sigma, 
-#'   K600_daily_sdlog}. \item If \code{pool_K600=='binned'} then 
+#'   K600_daily_sigma}. \item If \code{pool_K600=='binned'} then 
 #'   \code{K600_lnQ_nodes_centers, K600_lnQ_nodediffs_sdlog, 
-#'   K600_lnQ_nodes_meanlog, K600_lnQ_nodes_sdlog, K600_daily_sdlog}. \item If 
+#'   K600_lnQ_nodes_meanlog, K600_lnQ_nodes_sdlog, K600_daily_sigma}. \item If 
 #'   \code{err_obs_iid} then \code{err_obs_iid_sigma_scale}. \item If 
 #'   \code{err_proc_acor} then \code{err_proc_acor_phi_alpha, 
 #'   err_proc_acor_phi_beta, err_proc_acor_sigma_scale}. \item If 
@@ -144,8 +144,13 @@
 #'   dlnorm distribution for K600_daily, the daily rate of reaeration
 #' @param K600_daily_sdlog The lognormal scale parameter (standard deviation) of
 #'   a dlnorm distribution having meanlog equal to \code{K600_daily_meanlog} 
-#'   (when pool_K600 is 'none') or \code{K600_daily_pred} (otherwise) for 
-#'   K600_daily, the daily rate of reaeration
+#'   (when pool_K600 is 'none') or \code{K600_daily_predlog} (when pool_K600 is 
+#'   'normal') for K600_daily, the daily rate of reaeration as corrected for
+#'   temperature and the diffusivity of oxygen
+#' @param K600_daily_sigma The standard deviation of a dnorm distribution having
+#'   mean equal to \code{exp(K600_daily_predlog)} (applicable when pool_K600 is 
+#'   'linear' or 'binned') for K600_daily, the daily rate of reaeration as
+#'   corrected for temperature and the diffusivity of oxygen
 #'   
 #' @param K600_daily_meanlog_meanlog hyperparameter for pool_K600='normal'. The 
 #'   mean parameter (meanlog_meanlog) of a lognormal distribution of meanlog in 
@@ -336,7 +341,8 @@ specs <- function(
   K600_daily_meanlog_sdlog = 1,
   
   # hyperparameters for any K pooling or non-pooling strategy
-  K600_daily_sdlog = switch(mm_parse_name(model_name)$pool_K600, none=1, 0.2),
+  K600_daily_sdlog = switch(mm_parse_name(model_name)$pool_K600, none=1, normal=0.2, default=NA),
+  K600_daily_sigma = switch(mm_parse_name(model_name)$pool_K600, linear=10, binned=5, default=NA),
   
   # hyperparameters for hierarchical K600 - linear. defaults should be
   # reasonably constrained, not too wide
@@ -404,7 +410,7 @@ specs <- function(
   # daily simulation parameters
   discharge_daily = function(n, ...) rnorm(n, 20, 3),
   DO_mod_1 = NULL,
-  K600_daily = function(n, lnK600_daily_predlog=log(6), ...) rlnorm(n, lnK600_daily_predlog, 1),
+  K600_daily = function(n, K600_daily_predlog=log(10), ...) pmax(0, rnorm(n, K600_daily_predlog, 4)),
   GPP_daily = function(n, ...) pmax(0, rnorm(n, 8, 4)),
   Pmax = function(n, ...) pmax(0, rnorm(n, 10, 2)),
   alpha = function(n, ...) pmax(0, rnorm(n, 0.0001, 0.00002)),
@@ -468,8 +474,8 @@ specs <- function(
           features$pool_K600,
           none=c('K600_daily_meanlog', 'K600_daily_sdlog'),
           normal=c('K600_daily_meanlog_meanlog', 'K600_daily_meanlog_sdlog', 'K600_daily_sdlog'),
-          linear=c('lnK600_lnQ_intercept_mu', 'lnK600_lnQ_intercept_sigma', 'lnK600_lnQ_slope_mu', 'lnK600_lnQ_slope_sigma', 'K600_daily_sdlog'),
-          binned=c('K600_lnQ_nodediffs_sdlog', 'K600_lnQ_nodes_meanlog', 'K600_lnQ_nodes_sdlog', 'K600_daily_sdlog')),
+          linear=c('lnK600_lnQ_intercept_mu', 'lnK600_lnQ_intercept_sigma', 'lnK600_lnQ_slope_mu', 'lnK600_lnQ_slope_sigma', 'K600_daily_sigma'),
+          binned=c('K600_lnQ_nodediffs_sdlog', 'K600_lnQ_nodes_meanlog', 'K600_lnQ_nodes_sdlog', 'K600_daily_sigma')),
         if(features$err_obs_iid) 'err_obs_iid_sigma_scale',
         if(features$err_proc_acor) c('err_proc_acor_phi_alpha', 'err_proc_acor_phi_beta', 'err_proc_acor_sigma_scale'),
         if(features$err_proc_iid) 'err_proc_iid_sigma_scale'